backend/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
httpx
python-jose[cryptography]
passlib[bcrypt]
python-multipart
yfinance
openai
google-generativeai
psycopg2-binary
